feat: add season progress card
